refactor peaks and bin
