+++ amstrax/__init__.py
@@ -12,3 +12,4 @@ from . import hitfinder_thresholds
 from .hitfinder_thresholds import *
 
 from . import auto_processing
+
