+++ amstrax/plugins/peaks/__init__.py
@@ -1,2 +1,13 @@
-from . import peak_processing
-from .peak_processing import *
+from . import peaks
+from .peaks import *
+
+from . import peak_basics
+from .peak_basics import *
+
+from . import peak_classification
+from .peak_classification import *
+
+from . import peak_positions
+from .peak_positions import *
+
+
