+++ bin/process_run
@@ -1,8 +1,12 @@
+#!/usr/bin/env python
+
 import sys
-from amstrax.auto_processing import amstraxer
-from amstrax.auto_processing import amstraxer_easy
+import subprocess
+
+from amstrax.auto_processing import amstraxer, amstraxer_easy
 
 if __name__ == '__main__':
+
     args = amstraxer.parse_args()
 
     # Do import later to get fast --help
@@ -21,8 +25,14 @@ if __name__ == '__main__':
     run_collection.find_one_and_update({"number": int(run_name)},
                                        {"$set": {"processing_status": f"processing"}})
     print(f'Start amstraxer')
+
+
+    print(args)
+
+
     try:
         amstraxer_easy.main(args)
+
     except Exception as e:
         run_collection.find_one_and_update({"number": int(run_name)},
                                            {"$set": {"processing_status": f"failed due to {str(e)}"},
